+++ TIDALDL-PY/tidal_dl/gui_app/backend.py
@@ -410,7 +410,7 @@ class DemoBackend(TidekeeperBackend):
         SETTINGS.downloadVideos = False
         SETTINGS.multiThread = False
         SETTINGS.downloadDelay = True
-        SETTINGS.requestIntervalSeconds = 1.0
+        SETTINGS.requestIntervalSeconds = 3.0
         SETTINGS.saveAsFlac = False
         SETTINGS.usePlaylistFolder = True